adds RenderElementExtension, deprecate methods in SonataAdminExtension
