+++ src/Twig/Extension/SonataAdminExtension.php
@@ -19,17 +19,13 @@ use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\AdminBundle\Admin\FieldDescriptionInterface;
 use Sonata\AdminBundle\Admin\Pool;
 use Sonata\AdminBundle\Exception\NoValueException;
-use Sonata\AdminBundle\Templating\TemplateRegistryInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\DependencyInjection\Exception\ServiceCircularReferenceException;
-use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Translation\TranslatorInterface as LegacyTranslationInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Twig\Environment;
-use Twig\Error\LoaderError;
 use Twig\Extension\AbstractExtension;
 use Twig\Template;
 use Twig\TemplateWrapper;
@@ -62,6 +58,8 @@ class SonataAdminExtension extends AbstractExtension
     protected $pool;
 
     /**
+     * NEXT_MAJOR: Remove this property.
+     *
      * @var LoggerInterface
      */
     protected $logger;
@@ -81,6 +79,8 @@ class SonataAdminExtension extends AbstractExtension
     private $xEditableTypeMapping = [];
 
     /**
+     * NEXT_MAJOR: Remove this property.
+     *
      * @var ContainerInterface
      */
     private $templateRegistries;
@@ -121,13 +121,20 @@ class SonataAdminExtension extends AbstractExtension
     private $canonicalizeExtension;
 
     /**
+     * NEXT_MAJOR: Remove this property.
+     *
+     * @var RenderElementExtension|null
+     */
+    private $renderElementExtension;
+
+    /**
      * NEXT_MAJOR: Remove @internal tag, $translator & $securityChecker parameters and make propertyAccessor mandatory.
      *
      * @internal
      */
     public function __construct(
         Pool $pool,
-        ?LoggerInterface $logger = null,
+        ?LoggerInterface $logger = null, //NEXT_MAJOR: Remove this parameter
         $translator = null,
         ?ContainerInterface $templateRegistries = null,
         $propertyAccessorOrSecurityChecker = null,
@@ -213,33 +220,83 @@ class SonataAdminExtension extends AbstractExtension
     public function getFilters()
     {
         return [
+            //NEXT_MAJOR remove this filter
             new TwigFilter(
                 'render_list_element',
-                [$this, 'renderListElement'],
+                function (
+                    Environment $environment,
+                    $listElement,
+                    FieldDescriptionInterface $fieldDescription,
+                    $params = []
+                ) {
+                    return $this->renderListElement(
+                        $environment,
+                        $listElement,
+                        $fieldDescription,
+                        $params,
+                        'sonata_deprecation_mute'
+                    );
+                },
                 [
                     'is_safe' => ['html'],
                     'needs_environment' => true,
                 ]
             ),
+            //NEXT_MAJOR remove this filter
             new TwigFilter(
                 'render_view_element',
-                [$this, 'renderViewElement'],
+                function (
+                    Environment $environment,
+                    FieldDescriptionInterface $fieldDescription,
+                    $object
+                ) {
+                    return $this->renderViewElement(
+                        $environment,
+                        $fieldDescription,
+                        $object,
+                        'sonata_deprecation_mute'
+                    );
+                },
                 [
                     'is_safe' => ['html'],
                     'needs_environment' => true,
                 ]
             ),
+            //NEXT_MAJOR remove this filter
             new TwigFilter(
                 'render_view_element_compare',
-                [$this, 'renderViewElementCompare'],
+                function (
+                    Environment $environment,
+                    FieldDescriptionInterface $fieldDescription,
+                    $baseObject,
+                    $compareObject
+                ) {
+                    return $this->renderViewElementCompare(
+                        $environment,
+                        $fieldDescription,
+                        $baseObject,
+                        $compareObject,
+                        'sonata_deprecation_mute'
+                    );
+                },
                 [
                     'is_safe' => ['html'],
                     'needs_environment' => true,
                 ]
             ),
+            //NEXT_MAJOR remove this filter
             new TwigFilter(
                 'render_relation_element',
-                [$this, 'renderRelationElement']
+                function (
+                    $element,
+                    FieldDescriptionInterface $fieldDescription
+                ) {
+                    return $this->renderRelationElement(
+                        $element,
+                        $fieldDescription,
+                        'sonata_deprecation_mute'
+                    );
+                }
             ),
             new TwigFilter(
                 'sonata_urlsafeid',
@@ -290,12 +347,15 @@ class SonataAdminExtension extends AbstractExtension
     }
 
     /**
+     * NEXT_MAJOR: Remove this method
      * render a list element from the FieldDescription.
      *
      * @param object|array $listElement
      * @param array        $params
      *
      * @return string
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     public function renderListElement(
         Environment $environment,
@@ -303,22 +363,18 @@ class SonataAdminExtension extends AbstractExtension
         FieldDescriptionInterface $fieldDescription,
         $params = []
     ) {
-        $template = $this->getTemplate(
-            $fieldDescription,
-            // NEXT_MAJOR: Remove this line and use commented line below instead
-            $fieldDescription->getAdmin()->getTemplate('base_list_field'),
-            //$this->getTemplateRegistry($fieldDescription->getAdmin()->getCode())->getTemplate('base_list_field'),
-            $environment
-        );
+        if ('sonata_deprecation_mute' !== (\func_get_args()[4] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::renderListElement since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
+        }
 
-        [$object, $value] = $this->getObjectAndValueFromListElement($listElement, $fieldDescription);
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
+        }
 
-        return $this->render($fieldDescription, $template, array_merge($params, [
-            'admin' => $fieldDescription->getAdmin(),
-            'object' => $object,
-            'value' => $value,
-            'field_description' => $fieldDescription,
-        ]), $environment);
+        return $this->renderElementExtension->renderListElement($environment, $listElement, $fieldDescription, $params);
     }
 
     /**
@@ -393,54 +449,44 @@ class SonataAdminExtension extends AbstractExtension
     }
 
     /**
+     * NEXT_MAJOR: Remove this method
      * render a view element.
      *
      * @param object $object
      *
      * @return string
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     public function renderViewElement(
         Environment $environment,
         FieldDescriptionInterface $fieldDescription,
         $object
     ) {
-        $template = $this->getTemplate(
-            $fieldDescription,
-            '@SonataAdmin/CRUD/base_show_field.html.twig',
-            $environment
-        );
-
-        try {
-            $value = $fieldDescription->getValue($object);
-        } catch (NoValueException $e) {
-            // NEXT_MAJOR: Remove the try catch in order to throw the NoValueException.
-            @trigger_error(
-                sprintf(
-                    'Accessing a non existing value for the field "%s" is deprecated'
-                    .' since sonata-project/admin-bundle 3.67 and will throw an exception in 4.0.',
-                    $fieldDescription->getName(),
-                ),
-                E_USER_DEPRECATED
-            );
-
-            $value = null;
+        if ('sonata_deprecation_mute' !== (\func_get_args()[3] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::renderViewElement since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
         }
 
-        return $this->render($fieldDescription, $template, [
-            'field_description' => $fieldDescription,
-            'object' => $object,
-            'value' => $value,
-            'admin' => $fieldDescription->getAdmin(),
-        ], $environment);
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
+        }
+
+        return $this->renderElementExtension->renderViewElement($environment, $fieldDescription, $object);
     }
 
     /**
+     * NEXT_MAJOR: Remove this method
      * render a compared view element.
      *
      * @param mixed $baseObject
      * @param mixed $compareObject
      *
      * @return string
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     public function renderViewElementCompare(
         Environment $environment,
@@ -448,118 +494,44 @@ class SonataAdminExtension extends AbstractExtension
         $baseObject,
         $compareObject
     ) {
-        $template = $this->getTemplate(
-            $fieldDescription,
-            '@SonataAdmin/CRUD/base_show_field.html.twig',
-            $environment
-        );
-
-        try {
-            $baseValue = $fieldDescription->getValue($baseObject);
-        } catch (NoValueException $e) {
-            // NEXT_MAJOR: Remove the try catch in order to throw the NoValueException.
-            @trigger_error(
-                sprintf(
-                    'Accessing a non existing value for the field "%s" is deprecated'
-                    .' since sonata-project/admin-bundle 3.67 and will throw an exception in 4.0.',
-                    $fieldDescription->getName(),
-                ),
-                E_USER_DEPRECATED
-            );
-
-            $baseValue = null;
+        if ('sonata_deprecation_mute' !== (\func_get_args()[4] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::renderViewElementCompare since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
+        }
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
         }
 
-        try {
-            $compareValue = $fieldDescription->getValue($compareObject);
-        } catch (NoValueException $e) {
-            // NEXT_MAJOR: Remove the try catch in order to throw the NoValueException.
-            @trigger_error(
-                sprintf(
-                    'Accessing a non existing value for the field "%s" is deprecated'
-                    .' since sonata-project/admin-bundle 3.67 and will throw an exception in 4.0.',
-                    $fieldDescription->getName(),
-                ),
-                E_USER_DEPRECATED
-            );
-
-            $compareValue = null;
-        }
-
-        $baseValueOutput = $template->render([
-            'admin' => $fieldDescription->getAdmin(),
-            'field_description' => $fieldDescription,
-            'value' => $baseValue,
-            'object' => $baseObject,
-        ]);
-
-        $compareValueOutput = $template->render([
-            'field_description' => $fieldDescription,
-            'admin' => $fieldDescription->getAdmin(),
-            'value' => $compareValue,
-            'object' => $compareObject,
-        ]);
-
-        // Compare the rendered output of both objects by using the (possibly) overridden field block
-        $isDiff = $baseValueOutput !== $compareValueOutput;
-
-        return $this->render($fieldDescription, $template, [
-            'field_description' => $fieldDescription,
-            'value' => $baseValue,
-            'value_compare' => $compareValue,
-            'is_diff' => $isDiff,
-            'admin' => $fieldDescription->getAdmin(),
-            'object' => $baseObject,
-            'object_compare' => $compareObject,
-        ], $environment);
+        return $this->renderElementExtension->renderViewElementCompare($environment, $fieldDescription, $baseObject, $compareObject);
     }
 
     /**
+     * NEXT_MAJOR: Remove this method.
+     *
      * @param mixed $element
      *
      * @throws \RuntimeException
      *
      * @return mixed
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     public function renderRelationElement($element, FieldDescriptionInterface $fieldDescription)
     {
-        if (!\is_object($element)) {
-            return $element;
+        if ('sonata_deprecation_mute' !== (\func_get_args()[2] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::renderRelationElement since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
         }
 
-        $propertyPath = $fieldDescription->getOption('associated_property');
-
-        if (null === $propertyPath) {
-            // For BC kept associated_tostring option behavior
-            $method = $fieldDescription->getOption('associated_tostring');
-
-            if ($method) {
-                @trigger_error(
-                    'Option "associated_tostring" is deprecated since version 2.3 and will be removed in 4.0. Use "associated_property" instead.',
-                    E_USER_DEPRECATED
-                );
-            } else {
-                $method = '__toString';
-            }
-
-            if (!method_exists($element, $method)) {
-                throw new \RuntimeException(sprintf(
-                    'You must define an `associated_property` option or create a `%s::__toString` method'
-                    .' to the field option %s from service %s is ',
-                    \get_class($element),
-                    $fieldDescription->getName(),
-                    $fieldDescription->getAdmin()->getCode()
-                ));
-            }
-
-            return $element->{$method}();
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
         }
 
-        if (\is_callable($propertyPath)) {
-            return $propertyPath($element);
-        }
-
-        return $this->propertyAccessor->getValue($element, $propertyPath);
+        return $this->renderElementExtension->renderRelationElement($element, $fieldDescription);
     }
 
     /**
@@ -734,93 +706,56 @@ class SonataAdminExtension extends AbstractExtension
     }
 
     /**
+     * NEXT_MAJOR: Remove this method.
+     *
      * Get template.
      *
      * @param string $defaultTemplate
      *
      * @return TemplateWrapper
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     protected function getTemplate(
         FieldDescriptionInterface $fieldDescription,
         $defaultTemplate,
         Environment $environment
     ) {
-        $templateName = $fieldDescription->getTemplate() ?: $defaultTemplate;
-
-        try {
-            $template = $environment->load($templateName);
-        } catch (LoaderError $e) {
+        if ('sonata_deprecation_mute' !== (\func_get_args()[3] ?? null)) {
             @trigger_error(sprintf(
-                'Relying on default template loading on field template loading exception is deprecated since 3.1'
-                .' and will be removed in 4.0. A %s exception will be thrown instead',
-                LoaderError::class
+                'The %s method is deprecated in favor of RenderElementExtension::getTemplate since version 3.x and will be removed in 4.0.',
+                __METHOD__
             ), E_USER_DEPRECATED);
-            $template = $environment->load($defaultTemplate);
-
-            if (null !== $this->logger) {
-                $this->logger->warning(sprintf(
-                    'An error occured trying to load the template "%s" for the field "%s",'
-                    .' the default template "%s" was used instead.',
-                    $templateName,
-                    $fieldDescription->getFieldName(),
-                    $defaultTemplate
-                ), ['exception' => $e]);
-            }
         }
 
-        return $template;
+        return $this->renderElementExtension->getTemplate($fieldDescription, $defaultTemplate, $$environment);
     }
 
+    /**
+     * NEXT_MAJOR: Remove this method.
+     */
     private function render(
         FieldDescriptionInterface $fieldDescription,
         TemplateWrapper $template,
         array $parameters,
         Environment $environment
     ): string {
-        $content = $template->render($parameters);
-
-        if ($environment->isDebug()) {
-            $commentTemplate = <<<'EOT'
-
-<!-- START
-    fieldName: %s
-    template: %s
-    compiled template: %s
-    -->
-    %s
-<!-- END - fieldName: %s -->
-EOT;
-
-            return sprintf(
-                $commentTemplate,
-                $fieldDescription->getFieldName(),
-                $fieldDescription->getTemplate(),
-                $template->getSourceContext()->getName(),
-                $content,
-                $fieldDescription->getFieldName()
-            );
+        if ('sonata_deprecation_mute' !== (\func_get_args()[4] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::render since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
         }
 
-        return $content;
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
+        }
+
+        return $this->renderElementExtension->render($fieldDescription, $template, $parameters, $environment);
     }
 
     /**
-     * @throws ServiceCircularReferenceException
-     * @throws ServiceNotFoundException
-     */
-    private function getTemplateRegistry(string $adminCode): TemplateRegistryInterface
-    {
-        $serviceId = $adminCode.'.template_registry';
-        $templateRegistry = $this->templateRegistries->get($serviceId);
-
-        if ($templateRegistry instanceof TemplateRegistryInterface) {
-            return $templateRegistry;
-        }
-
-        throw new ServiceNotFoundException($serviceId);
-    }
-
-    /**
+     * NEXT_MAJOR: Remove this method
      * Extracts the object and requested value from the $listElement.
      *
      * @param object|array $listElement
@@ -828,43 +763,24 @@ EOT;
      * @throws \TypeError when $listElement is not an object or an array with an object on offset 0
      *
      * @return array An array containing object and value
+     *
+     * @deprecated since sonata-project/admin-bundle 3.x and will be removed in 4.0
      */
     private function getObjectAndValueFromListElement(
         $listElement,
         FieldDescriptionInterface $fieldDescription
     ): array {
-        if (\is_object($listElement)) {
-            $object = $listElement;
-        } elseif (\is_array($listElement)) {
-            if (!isset($listElement[0]) || !\is_object($listElement[0])) {
-                throw new \TypeError(sprintf('If argument 1 passed to %s() is an array it must contain an object at offset 0.', __METHOD__));
-            }
-
-            $object = $listElement[0];
-        } else {
-            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an object or an array, %s given.', __METHOD__, \gettype($listElement)));
+        if ('sonata_deprecation_mute' !== (\func_get_args()[2] ?? null)) {
+            @trigger_error(sprintf(
+                'The %s method is deprecated in favor of RenderElementExtension::getObjectAndValueFromListElement since version 3.x and will be removed in 4.0.',
+                __METHOD__
+            ), E_USER_DEPRECATED);
         }
 
-        if (\is_array($listElement) && \array_key_exists($fieldDescription->getName(), $listElement)) {
-            $value = $listElement[$fieldDescription->getName()];
-        } else {
-            try {
-                $value = $fieldDescription->getValue($object);
-            } catch (NoValueException $e) {
-                // NEXT_MAJOR: throw the NoValueException.
-                @trigger_error(
-                    sprintf(
-                        'Accessing a non existing value for the field "%s" is deprecated'
-                        .' since sonata-project/admin-bundle 3.67 and will throw an exception in 4.0.',
-                        $fieldDescription->getName(),
-                    ),
-                    E_USER_DEPRECATED
-                );
-
-                $value = null;
-            }
+        if (null === $this->renderElementExtension) {
+            $this->renderElementExtension = new RenderElementExtension($this->propertyAccessor, $this->templateRegistries, $this->logger);
         }
 
-        return [$object, $value];
+        return $this->renderElementExtension->getObjectAndValueFromListElement($listElement, $fieldDescription);
     }
 }
